requirements.txt
# Core dependencies
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# MCP dependencies
mcp>=0.2.0

# Agent dependencies
pydantic-ai>=0.4.0
openai>=1.0.0

# Optional dependencies
logfire>=0.15.0  # For enhanced logging (optional)
